Major rework of beam and baseline fitting
Instead of a joint non-linear least-squares optimisation, the Gaussian beam
and 2-D polynomial baseline are alternately fitted to the data in an iterative
process. The final step allows refinement of the beam by refitting it only to
the points close to the peak of the beam pattern.

The fitted beam and baseline are stored as separate objects in the compound
scan. The BeamPatternFit object is a thin veneer on a GaussianFit that
presents friendlier terminology and units (center/width instead of mean/var).

Also fixed formula for expected beamwidth.

git-svn-id: https://katfs/svnDS/code/scape/trunk@1666 b56c7a7d-ac47-0410-a905-b6f7de78896a