Use same beamwidth for x and y
Beam pattern fit now uses Gaussian with a single shared variance for both
target coordinates, as the beam pattern is expected to be very close to
circular in cross-section. It there has the same width in x and y coords.
This will especially improve the fit in the case of horizontal-only or
vertical-only scans across a source, as the width in the unscanned
direction is typically poorly constrained.

git-svn-id: https://katfs/svnDS/code/scape/trunk@1685 b56c7a7d-ac47-0410-a905-b6f7de78896a