Next incarnation of beam + baseline fitting
Instead of a baseline per compound scan, we now have a baseline per scan.
This collection of (linear) baselines are still initialised via knowledge of
the entire compound scan, though, to improve fitting.

The advantage of per-scan baselines is that it can eliminate short-term
glitches that affect a small set of scans, as is abundant in the XDM data.
The beam is still done globally, across the entire compound scan.

Furthermore, some scans will not have baselines, as they are too far away
from the beam center. The sole purpose of baselines is therefore to improve
the beam fit (especially its amplitude). The meaning of baselines for
Stokes Q, U and V still has to be explored.

git-svn-id: https://katfs/svnDS/code/scape/trunk@1953 b56c7a7d-ac47-0410-a905-b6f7de78896a