Allow different beamwidths in x and y
Allow fitted beam to have different beamwidths in x and y directions.
This is useful when fitting XX and YY beams, which usually are slightly
elliptical.

Also remove 1.0047 beam height correction factor, as this seems
counterproductive on GPS and Orion A beams. On perfect Gaussian beams,
the fitting routine tends to underestimate the beam height, while on real
data, the beam height is overestimated. By removing this correction, the
real data fit is improved.

git-svn-id: https://katfs/svnDS/code/scape/trunk@3036 b56c7a7d-ac47-0410-a905-b6f7de78896a