Refactor fitting routines
Cleaned up exception handling, replacing inappropriate usage of
AttributeError, TypeError and KeyError with a user-defined NotFittedError
and ValueError otherwise.

Unified the calling convention for ScatterFit .fit and __call__ methods,
so that a sequence of D-dimensional x input vectors are stored as *column*
vectors in an array of shape (D, N). Previously, NonLinearLeastSquaresFit
and GaussianFit accepted row vectors for x instead (while the rest did
column vectors), as this was simpler to vectorise a function for this setup.

However, column vectors have an important notational advantage: 2-D fitting
can be written as interp.fit([x, y], z), given 1-D input vectors x, y and z.
This corresponds closely to the way 2-D fitters are already called in SciPy.

git-svn-id: https://katfs/svnDS/code/scape/trunk@3571 b56c7a7d-ac47-0410-a905-b6f7de78896a