Add uncertainty to beam fitting
Major rewrite of GaussianFit in order to incorporate uncertainty.

This now uses the standard deviation as parameter (instead of the variance),
as it simplifies calculations and is more user-friendly. Added the proper
Jacobian for a diagonal Gaussian function to the non-linear least-squares fit
(instead of estimating it from the data), as this is essential for covariance
estimation. The fit member now accepts a std_y parameter.

Fixed bug in NonLinearLeastSquaresFit where Jacobian evaluation ignored std_y.
Added unit test to check covariance matrix estimation in
NonLinearLeastSquaresFit. Fixed test bug where parameters where inadvertently
initialised to integers (resulting in slow and wrong integer optimisation).
Improved test docstrings.

git-svn-id: https://katfs/svnDS/code/scape/trunk@8480 b56c7a7d-ac47-0410-a905-b6f7de78896a