Refactor of beam/baseline fitting
The underlying function is now agnostic of DataSet/CompoundScan/Scan and
operates purely on arrays of data. This makes it much easier to unit test
and to extend. The rest of the functionality (spike removal, DOF estimation,
beamwidth estimation and other dataset-specific things) is now concentrated
in a CompoundScan method, which is its natural home. The high-level DataSet
method is mostly superfluous and now simply calls the CompoundScan method on
all its compscans. This was a leftover from an era when the lower-level scans
and compound scans could not access information from higher levels (since
rectified), which left the DataSet as the only place to do high-level
decisions. Also added the start of uncertainty handling, but this is still
disabled. Passes all previous unit tests.

git-svn-id: https://katfs/svnDS/code/scape/trunk@8499 b56c7a7d-ac47-0410-a905-b6f7de78896a