Handle existing pointing adjustments in fitting
This is especially relevant when verifying that reference pointing
works. Now apply any existing pointing adjustment as if it is part of
the existing pointing model. This results in two extra numbers in the
pointing_offsets output array: relative (az, el) adjustment on top of
both the existing pointing model and the existing adjustment. These
numbers should be small on the second run of this script on the same
target, which allows for simple verification. Also only log the relative
adjustment values to avoid confusing the user. Improve the docs, which
now also distinguishes between pointing "offsets" and pointing
"adjustments" (the former includes the PM).