Ensure we get the gains of the final offset
The current version of the script keeps tracking the final offset and
the gain solutions then end up outside the expected time range and
get ignored. One way to terminate the solution interval is to stop
the antennas (as in bf_phaseup). This is quite harsh though, and a
softer solution is to clear the pointing offsets which take the dishes
back to the main target. We assume that the final offset is not the
origin - it wouldn't be if we stick to the usual raster patterns.
Also remember to wait for the gains to materialise (use bf_phaseup as
a guide, although it is currently worst-case).