Consolidate normalisation in bp_fit
The original bandpass fitter always centres the phase on zero and also
optionally normalises the amplitude to an average of one (both as a
function of frequency). The pipeline uses the default normalisation
option of True, so both steps are typically performed for bpcal tags.

Some users prefer an unnormalised version of bp_fit, so rather combine
the phase normalisation with the amplitude normalisation so that both
are controlled by the same option. This has the added benefit of being
faster, as the normalisation factor can be computed and then broadcast
and applied once to the full set of gains.