Debug and refactor coarse delay calculation
The large delay errors described in MKAIV-865 are due to a bug in the
conversion of the index of maximum FFT magnitude to a frequency.
Only the "good" data is transformed with flagged channels stripped out
(a second bug yet to be addressed) but the index is FFT-shifted based
on the full number of channels (`num_chans`). Once more than 50% of
the channels are flagged, large indices are not mapped to small
negative frequencies anymore and remain as large delays.

Remove the homegrown conversion and use np.fft.fftfreq instead.
This removes the need for `num_chans` entirely. Also note that the
coarse delay step assumes that the channel frequencies are regularly
spaced - only the difference between the first two channels matter.

Finally, do a small bit of refactoring to the coarse delay extraction
to reduce the amount of code.

This addresses JIRA ticket MKAIV-865.