Improve handling of NaNs in delaycal
The existing scheme stripped out any channel that contained even a
single NaN. The more aggressive flagging heralded by the 2-D flagger
occasionally causes all channels to be discarded, leading to the
infamous "Invalid number of FFT data points (0) specified" crash.

This strategy also causes errors in the coarse delay estimation as
the relationship between channels are disrupted by the excision, which
broadens the spectral peak associated with the delay and may introduce
spurious sidelobes. The channel excision results in disjoint chunks
of a single sinusoid and fewer data points.

The proposed improvement is to zero the NaNs instead of excising them.
This also broadens the spectral peak and may introduce spurious
sidelobes but at least guarantees a successful FFT. In addition, the
remaining good data in each channel is preserved and not chucked out
with the NaN. The effect of the NaN mask is similar to that of a dirty
image of a point source.

Also refactor the coarse delay correction to simplify the code.

This addresses JIRA tickets MKAIV-865, MKAIV-779 and SR-804.