Generate additional diagnostic plots in the cal report.
In order to assess the success of the calibration procedure the report now includes a number of plots of calibrated data.

At the end of the processing of each scan, calibration solutions are applied and the data are averaged per scan and across frequency to 1024 channels. The corrected data for every scan is stored in a dictionary containing arrays of calibrated data and upon receiving a StopObservation event this dictionary is passed to the report writer.

The report generates the following additional plots:
* Elevation vs time for the reference antenna
* % of data flagged during calibration
* Amp and Phase vs frequency for calibrators.
* Amp and Phase vs time for phase calibrated calibrators. ( Phase will not be aligned for non-phase calibrated calibrators.)
* Amp and Phase vs uv distance for calibrators
* Amp vs frequency for target fields
* Amp vs uv distance for target fields