Improve performance of calprocs_dask.wavg_full_f
The old version was generating a chunk per output along the frequency
dimension, which meant thousands of chunks and poor efficiency. The new
version rechunks if necessary to force alignment, then delegates to the
non-dask version.

Also fixes SR-1034 by updating constructs marked deprecated in dask
0.16, and enforcing 0.16 as a minimum version.

The tests for wavg_full_t just compare the results against the non-dask
version, which does not (yet) have a test of its own.