Style and Docstring Cleanup
Other miscellaneous changes:
Also fix order of pylab import in report.py
Fix the order of the scans in the corrected data dictionary. Previously the dictionary was updated with np.append after every processed track, but this reverses the order of the scans as the pipeline processes scans in reverse order.