Switch to using dask.distributed
It seems to make better scheduling decisions (see dask/dask#3139), and
has some on-line performance analysis tools.

The old --dask-diagnostics option is repurposed to take a port number.
To avoid creating port conflicts with other code (particularly when run
under a version of master controller that hasn't been updated), the port
number must be explicitly specified - the default is not to provide the
diagnostics.