Handle the 'product_' part of key in session
This makes the scripts nicer, which can now refer to gain solutions
as 'G' instead of 'product_G'. In effect, the stream is called
'cal_product'. We are already hiding the 'cal_' part, so this is a
natural progression. This tames the monster of 'product_KCROSS_DIODE'
with its curious mixed case.