Let k_fit return NaN if no valid data is available
If all the data of a single receptor is flagged, the residual delay
fit will bomb out with a "LinAlgError: Arrays cannot be empty" error.
The coarse fit soldiers on with zeros instead of NaNs via good_pol_data
but the subsequent bandpass fit will restore NaNs for totally flagged
data. In this case, do not attempt to fit a phase slope and return
a delay value of NaN instead.

This addresses JIRA ticket MKAIV-1096.