Reduce memory-footprint of 'n_flags'
Previously the cal-report produced rfi summary plots from the `n_flags` array
in the dictionary of corrected data passed to the report. This array has a size of (n_tracks x n_channels x n_pols x n_baselines) which grew unreasonably large for observations with many tracks.

To mitigate this the `n_flags` array has been replaced with two smaller arrays:
* `t_flags` with finaly size (1024 x n_pols x n_baselines)
* `bl_flags` with size (n_tracks x 1024 x n_pols)

The `t_flags` array represents the sum of all flags over time and a cumulative sum is calculated at the end of every run of the pipeline.

While `bl_flags` represents the per scan average fraction of flagged baselines. Each per scan output is stored in a list with all outputs concatenated into an array at the end of the observation.