Determine the reference antenna based on the first scan of data
Use the data in the first scan after the start of the subarray to determine the reference antenna. Once the refant is set the same refant is used for the rest of the subarray, no further attempt to determine its quality is made ( although the hope is to improve this in a later pull request).

The refant is chosen based on a procedure similar to that used to determine the coarse delay of the data. It uses the height of the peak of the fft of the visibilities to a particular antenna as a quality/SNR indicator, with a stronger peak implying a better quality. As each antenna will have n_ant-1 baselines to it the sum of the peak height across these n_ant-1 baselines is used to measure quality. The antenna with the largest sum is used as the refant.

The refant is chosen after flagging the data, this should ensure that heavily flagged antennas are unlikely to be chosen as the refant.

The refant selection procedure uses that same region of the spectrum as the delay calibration.