Support multiple flag streams
To deal with configuration of multiple streams, each of which has
multiple settings, the command-line options are now all packed into a
JSON document (but when passed from telstate, can be the Python object
structure directly).

The schema includes the option to specify continuum averaging of the
flag stream, but it is not yet implemented.

Relates to SR-1351.