Normalise bandpass after flagging
This improves the normalisation of the phases of the bandpass.

This commit also fixes a bug in the normalisation of the amplitudes. Previously
the mean amplitude was incorrectly determined by dividing by the total number
of channels in the bandpass, this has been corrected to divide by the
total number of non-NaN channels in the bandpass.