Improve bandpass normalisation
In the current split-cal implementation for 32K channels
the bandpass is divided into 4 sections and each section if solved for
and normalised independently. The independent normalisation of each of these
regions results in a bandpass solution which has small discontinuities at the
edges of the sections.

To ensure the normalised bandpass solution is continuous everywhere this commit
uses only a small region of the band to calculate a per antenna normalisation factor which
normalises the bandpass solution within this region. This factor is then applied to the 4
sections of the unnormalised bandpass solutions before storing them in the solution store and telstate.

The normalisation factor is calculated from the same region of the band that is used to
calculate the `G` solutions.