Add aiomonitor to cal
The main script got shuffled around somewhat to accommodate the annoying
requirement that aiomonitor has to be started before the event loop is
run. While that would have been simple to do, it would not have given
access to all the interesting local variables. So instead, most of the
script code is now run before starting the event loop (in `main`), and
only the minimum code necessary is run inside the event loop (in `run`).

Also fixed in bug in the `__exit__` for CalDeviceSender, where the
sender process was not killed.