Fix reference pointing to use obs_activity
The script predates the array activity sensor. Because it directly
drove the dishes and slept, it did not set the activity, resulting
in a single big 'track' for the entire observation. Rather use
session.track(), which has the added advantage of respecting the
receptor quorum. Also update the log messages to avoid duplication
and provide useful info.

This has been verified to work on site (see SBID 20190522-0028).

This addresses JIRA ticket SR-1807.