Calculate SNR for K, B, and G solutions
Signal to Noise Ratios (SNR) per antenna for these solutions are estimated
as 1/rms of the phase of the residuals after applying the solutions.

As the rms is calculated across all baselines to a given antenna a single poor
baseline can bias the SNR estimate to a very low value. To guard against this antennas
with poor rms estimates (rms > 0.2) on 80% of their baselines are excluded when
estimating the SNR of other antennas.