Make CalSolution and CalSolutions target aware
Flux calibration requires that the gains for a given gain calibrator
which were scaled to a flux density of 1.0 Jy are re-scaled by the measured
flux density of said gain calibrator. This scaling requires measuring the
flux density of the gain calibrators by computing the ratio
of the (flux calibrated) gains of the flux calibrators to the
average (uncalibrated) gains for a whole observation on each gain
calibrator. The scaling factor will be different for each gain calibrator
in the observation so we need to retrieve the gains on a particular
target independently from the solution store at the end of an observation.

Therefore, store target information in CalSolution and CalSolutions.
A target name is required for a CalSolution, but is optional for a
CalSolutions object. This is because CalSolutions might be interpolated
from multiple targets. I've added a new method 'get_target' to
CalSolutionStore which retrieves all of the gains for a specific
target and stores these in a CalSolutions object. Also tests.