Add function to compute the flux density of gain calibrators.
The docstring for F_cal in calprocs.py contains the relevant info
about how this is done. F_cal is run in control.py after an
ObservationEndEvent. The computed flux densities and their
associated errors are stored in telstate with keys 'cal_CBID_measured_flux'
and 'cal_CBID_measured_flux_STD'.

Also scale the gains on the gain calibrator and store these in a new
solution store called G_SCALE. This is used when applying the gains
to ensure the report contains scaled gains. It is set up such that
the solution store G_FLUX is used for 'bpcal' targets which have a flux
model. All other targets get their solution from 'G_SCALE' either
interpolated to the target or from the target itself if 'G_SCALE'
has gains for the target.