Make cal work with latest version of dask
There are two fixes:

1. A straight-up bug where a dependency set was made a string instead of
a set of strings. I'm not sure how this ever worked!
2. A workaround for an old dask bug (dask/dask#3139) caused us to run
into a new dask bug (dask/dask#5850). I've removed the workaround,
although we should test on a real-size problem to be sure that the
workaround is no longer needed.

I've also bumped the version to the latest to ensure that we pick up
whatever fixed the old bug.