Choose a new reference antenna if necessary
This is to address https://skaafrica.atlassian.net/browse/SR-1228.
If the reference antenna selected by the pipeline at the beginning of
the subarray is flagged for any reason it results in NaN solutions. This
commit allows the pipeline to select a new refant at the start of a
new capture-block if the previously selected refant is flagged.

This commit allows the pipeline to select a new refant if:
* one hasn't already been selected (e.g at the start of the subarray)
* the current refant has a flag fraction >80%