Allow non-stationary flux, noise var and gains
This makes the data more realistic. The source flux varies as a function
of frequency, as does the SEFD. In addition, multiply the resultant
sum with a frequency-dependent gain. Some channels are zeroed to
indicate flagged data (use an actual RFI mask from 1556574656).
Rework the CRB to cater for flagged and non-stationary data. It seems
to work well :-) Add an experiment to test this.