Speed up fft_secant and limit iterations
Only calculate derivatives for frequencies that have not converged yet.
This speeds up the final iterations a lot, and helps when a handful of
frequencies are taking longer to converge that the rest. Overall I got
about a 2x speedup of `fft_secant`.

In the process, negate the mask used to select which frequencies to
work on (i.e. use `active = ~done`).

Also limit the number of iterations to 100, since the secant method is
not guaranteed to converge. Doing baseline cal prep revealed some input
data for which it never converges, which is not nice.