Estimate estimator variance in general case
Come up with an alternative formula for the Cramer-Rao bound (hopefully)
in the general case of missing data points. I speculated that the
(N * N - 1) term has to do with the extent of the data points, which is
why I initially had it as (N_range ** 2 - 1). The frequency variance is
related to the curvature around the spectral peak. This is the second
derivative of the spectrum, which turns into the second moment of the
data points via the Fourier transform. If you have N regularly-spaced
data points, the second central moment is (N * N - 1) / 12, which
also explains the factor 6 in the regular CRB :-)

This again closely matches the fft_secant results, so there is some
hope that it is correct.