Split out data generation function
This is in preparation for antenna-based delays. Also change the
nomenclature from "frequency" estimation to "phase slope" estimation
(in units of radians/channel), which is much clearer. Generate a
sequence of slopes outside `_generate_data`, since we need to construct
it specially for antenna-based delays.

Change the default channel count to 1024 to speed things up.