Return antenna-based and baseline-based results
When doing antenna-based experiments, express the residual delays
either in antenna or baseline form. Choose the appropriate version with
a parameter. Clean up the error measurement by calculating the CRB
separately, and returning the residuals instead of standard deviations.
This allows for more fine-grained analysis of the errors. Use a more
robust IQR-based standard deviation to ignore large errors for now.