Mitigate rogue amplitude spikes in data
While the FFT-based delay estimators are generally tolerant of variable
signal amplitudes, they can choke on massive amplitude spikes. This
distorts the peak in the magnitude spectrum until it drops below the
background, causing catastrophic failure.

Test for this by adding a rogue spike to the generated data. Mitigate
this up front for all methods by killing all samples with an amplitude
above a certain factor of the median amplitude. This is much better than
clipping the amplitude, especially for high flux / SNR scenarios, where
clipping causes the error to remain stuck at the error level of a 10-Jy
calibrator.

The funny thing is that the secant method still attains the CRB, even
if the spike threshold is equal to the median itself, thereby killing
half the samples... (in a random fashion though -- CS to the rescue?)