Improved Mattieu2 algorithm
This now matches FFT-Secant nearly perfectly. Mattieu ascribes this to
improved initialisation:

"The reason why it is important for you to double sampling is because
the nearest sample may be part of the sidelobe instead of the mainlobe
of the sinc. But I now found a trick that I do not have to double
sample to resolve the issue. Newer version - all I do is I change the
starting point to be closer by half (instead of measuring again what I
already knew from the initial FFT)."