SPR1-2965 Fix basic package metadata
Move main package metadata from setup.py to the modern pyproject.toml
(see PEP 621).

Add a README, using GitHub Flavored Markdown (.md) instead of
reStructuredText (.rst) since PyPI accepts it these days.

Switch from katversion to setuptools_scm to determine the package
version dynamically. I've also considered dunamai, which

  - Allows branch names in local version strings (useful for debugging)
  - Supports the poetry build system (but we are not using that yet)

but unfortunately it

  - Needs setup.py outside of poetry (a bit of a showstopper currently)
  - Is not part of the Python Packaging Authority (PyPA) family

Pick a version of setuptools_scm that works with Python 3.6 for now.
Bump the Python requirement accordingly (it's probably not tested on
Python 3.5 anyway).

Strip down the list of dependencies from the full katsdpcal list to
what's needed for this package. Limit katdal and katpoint to the 0.x
series. Keep a minimal setup.py for editable installs (but eventually
this could be removed too). And set the status at least to "Beta"
(and hopefully to "Stable" once the dust settles :-) ).