SPR1-1964 Improve reference antenna selection
Modify the refant selection to select the longest baseline antennas
after excluding those with very low peak-to-noise ratios in the FFT of
their delay spectrums. Exclusion is based on being lower than 1 MAD
from the median value for all antennas.

The median absolute deviation is calculated by a function introduced
in scipy 1.5.0.