SPR1-3119 Reproduce Condon simulation as unit test
Before we start ripping out the `scikits.fitting` fitter, let's first
check how well it works so that we don't regress... Reproduce the
simulation in Condon's classic "Errors in Elliptical Gaussian Fits"
paper with some minor modifications:
  - Use wider Gaussians to avoid fitting failures (FWHM >= 5 pixels).
  - Dither both the location and width of the Gaussian beams to make
    it more realistic.
This basically recreates the data that underpins the histograms in
Figure 1 in the paper.

Do some normality tests on the parameter errors as a unit test, which
also verifies that our fitter achieves the desired accuracy.

Also check the parameter uncertainties returned by the fitter against
the theoretical uncertainties, with quite a bit of thumb sucking to
estimate the variance of the uncertainty estimate...