SPR1-3119 Remove scikits.fitting dependency
Replace GaussianFit with the very similar Gaussian2D model in Astropy's
modeling module. This is an existing dependency with better support and
more features. It also avoids having to add scikits.fitting to the
docker base image in order to satisfy katsdpcal's dependencies due to
the way install_pinned.py works. For the record, scikits.fitting started
out in the scape package back in 2007, while Astropy models had its
initial commit only in 2013, in case you wondered why we did not use
Astropy models from the start. :-)

Both rely on the non-linear least-squares solvers in `scipy.optimize`.
GaussianFit uses the older `leastsq` routine. We have now opted for the
recommended `LMLSQFitter` based on the newer `least_squares` routine.
It is a good compromise between speed and robustness.

For the record, `scikits.fitting` is still faster since its solver is
customised for "diagonal" (axis-aligned) Gaussians. Timing results
obtained with `pytest tests/test_pointing.py --durations=1`:

  - GaussianFit (leastsq): 2.0 s
  - Gaussian2D + LevMarLSQFitter (leastsq): 4.0 s
  - Gaussian2D + LMLSQFitter (least_squares): 4.7 s
  - Gaussian2D + TRFLSQFitter (least_squares): 7.0 s

The `LevMarLSQFitter` has some extra fitting failures which counted
against it. Reduce the number of Gaussians to fit to make up for the
slower test.

The new `BeamPatternFit` also achieves the desired accuracy on the
parameter fits, but the uncertainty of the uncertainty is slightly
higher (maybe because it's more generic underneath?).

Replace the somewhat superfluous FWHM helper functions with:

  - fwhm_to_sigma(width) -> width / GAUSSIAN_SIGMA_TO_FWHM
  - sigma_to_fwhm(sigma) -> sigma * GAUSSIAN_SIGMA_TO_FWHM