SPP1-39 Ensure wavg_t_f has the correct dimensions
The function `calprocs_dask.wavg_t_f` averages over all times and in frequency blocks.
It calls `calprocs_dask.wavg_f` which manually aligns the dask chunks to the frequency
blocks that are being averaged over.

However this manual alignment was being done on axis=1, which is the
incorrect axis for the case when the time axis has been averaged away
and instead the alignment should be done on axis=-3 to be general for
the case where a time axis does and doesn't exist.

Misalignment between the chunk sizes and the dimensions of the final
averaged array resulted in averaged arrays with the incorrect dimensions
which caused problems when assembling the arrays in the report which
assumes arrays of the appropriate dimensions.